pyproject.toml
[project]
name = "dune-monorepo"
version = "0.1.0"
description = "DUNE APA winder control and wire tension measurement"
requires-python = ">=3.12"
dependencies = [
    # dune_winder
    "pycomm3>=1.2.14",
    "influxdb-client>=1.40.0",
    # dune_tension
    "flask>=3.1.0",
    "numpy>=1.26.4",
    "matplotlib>=3.10.3",
    "nnaudio>=0.3.3",
    "onnxruntime>=1.16.0",
    "pandas>=2.3.0",
    "pesto-pitch>=2.0.1",
    "pyserial>=3.5",
    "requests>=2.32.3",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "sounddevice>=0.5.2",
    "ty>=0.0.32",
    "scikit-learn>=1.8.0",
]

[project.scripts]
# dune_winder
dune-winder = "dune_winder.main:main"
plc-rung-transform = "dune_winder.plc_rung_transform:main"
plc-import = "dune_winder.plc_import:main"
plc-manifest = "dune_winder.plc_manifest:main"
plc-sync = "dune_winder.plc_sync:main"
plc-convert-rllscrap = "dune_winder.convert_plc_rllscrap:main"
# dune_tension
dune-tension-gui = "dune_tension.main:main"
dune-tension-periodic-plots = "dune_tension.periodic_save_plots:main"
dune-tension-download-m2m = "dune_tension.download_tension_data:main"
dune-tension-plot-sqlite = "dune_tension.plot_tension_from_sqlite:main"
dune-tension-stream-sweep = "dune_tension.streaming.cli:main_stream_sweep"
dune-tension-stream-rescue = "dune_tension.streaming.cli:main_stream_rescue"
dune-tension-fit-focus = "dune_tension.streaming.cli:main_fit_focus"
dune-tension-stream-replay = "dune_tension.streaming.cli:main_stream_replay"
dune-tension-average-profile-clouds = "dune_tension.average_profile_clouds:main"
dune-tension-average-profile-clouds-gui = "dune_tension.gui.average_profile_clouds_app:main"
dune-spectrum-compare = "spectrum_analysis.compare_pitch_cli:main"
dune-spectrum-scroller = "spectrum_analysis.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dune_tension = ["*.json", "*.npy", "*.npz"]
dune_winder = ["version.xml", "py.typed"]
spectrum_analysis = ["*.json"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.9.9",
]

[tool.pytest.ini_options]
testpaths = ["tests/dune_tension", "tests/dune_winder"]
addopts = "-m 'not ladder_sim'"
markers = ["ladder_sim: ladder simulator tests (not in active development)"]

[tool.ruff]
indent-width = 4
extend-exclude = ["src/dune_winder/manual_port.py"]

[tool.ruff.lint]
ignore = ["F401", "F811", "F841", "E741"]

[tool.ruff.lint.per-file-ignores]
"tests/dune_tension/*.py" = ["E402"]
